requirements.txt
certifi==2020.12.5
chardet==3.0.4
cos-python-sdk-v5==1.9.0
crcmod==1.7
dicttoxml==1.7.4
idna==2.10
natsort==7.1.0
Pillow==8.0.1
pycryptodome==3.9.9
PyMySQL==0.10.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
